Improve back to top link accessibility
Move the back to top link after the main content
---
This means that anyone using a keyboard can access the back to top link
if they're navigating to the bottom, whereas perviously the link was
just after the navigation.

Moving this forces us to have to change our approach having the back to
top link follow the user.
We cannot use `position: sticky`, so have to make more use of
intersection observers and implement the desired behaviour ourself.

Move the link to the bottom of the page visually
---
For users with motor impariments this may be easier to access as it's
closer to the bottom of the page.
This change also means that the fallback for older browsers and
JavaScript being disabled is the link at the bottom of the page.