Adjust Nginx auth config for __canary__
The nginx config previously set auth_basic to 'off' for /__canary__/ with the trailing slash.
This meant /__canary__ required basic auth, but /__canary__/ didn't.

This is easy to miss/mix up (e.g: when changing or setting up new healthchecks). Instead, we can change the nginx config to set
auth_basic to 'off' for __canary__ (minus the trailing slash) and this should apply to both __canary__ and __canary__/

http://nginx.org/en/docs/http/ngx_http_charset_module.html#charset_types